Revert "Модуляция"
